add plan
```
Look carefully at instructions.txt and in the resources/ directory. In particular, note the resources/sign/ and resources/id/ pair of example applications

Expand instructions.txt into a detailed implementation plan. The implementor will be starting from a copy of the id/ app as a template

Note in particular that bindings for the UI will be generated when the app is built with `kit build --hyperapp`. As such, first design and implement the backend; from the backend, the interface will be generated; finally design and implement the frontend to consume the interface. Subsequent changes to the interface must follow this pattern as well: start in backend, generated interface, finish in frontend
```

```
edit the plan. Do NOT create the API. The API is machine generated. You create types that end up in the API by defining and using them in functions in the Rust backend "hyperapp"
```

```
almost there. The #[http] header does not take any arguments. You can match on the path inside the function body with `hyperware_app_common::get_path()` if needed
```

```
process_lib/src/vfs/directory.rs                                                                                                                                                                      │
[Pasted text #1 +97 lines]

process_lib/src/vfs/file.rs                                                                                                                                                                           │
[Pasted text #2 +384 lines]                                                                                                                                                                           │

process_lib::vfs
[Pasted text #3 +160 lines]

get the most useful function signatures from here and make brief
descriptions of what they do/how to use them & put them in the plan
```